requirements.txt
smbus2
python-telegram-bot